Rename global worker
