+++ lib/renovate.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 
-const renovateWorker = require('./workers/index');
+const globalWorker = require('./workers/global');
 
-renovateWorker.start();
+globalWorker.start();
