+++ lib/workers/global.js
